requirements/production.txt
# Production dependencies
-r base.txt

# Production WSGI Server
gunicorn==23.0.0

# Production Database Connection Pooling
django-db-connection-pool==1.0.0

# Security
django-security==0.20.0

# Monitoring
django-prometheus==2.3.1

# Health Checks
django-health-check==3.18.1
